Modified the "lopping" procedure for determining the fraction cell height.
The original formulation first filter the model depth using
the hFacMin parameter. Any fractional cells less than hFacMin
were rounded to the nearer of 0 or hFacMin. Then the depths were
filtered using hFacMinDr where the actual size was tested.
This meant that some cells could be rounded twice and also
that if hFacMinDr was specified as larger than the nominal
thickness that the fractional cell thickness could end up
larger than 1. Clearly a bad idea (?) but to cause this one
would have to specify the hFacMin and hFacMinDr parameters in
a inconsistent way.

The new method combines hFacMin and hFacMinDr first and then
does a single filtering of model depths. If hFacMinDr is
greater than the nominal layer thickness, the later is used
and the fractional cell thicknesses are set to 1. Otherwise,
layers thicknesses are rounded to using the larger of
hFacMin and hFacMinDr/drF(K).

Using the new method, setting hFacMin=0 and hFacMin=100
means that lopping only occurs for cell thicker than 100
and the minimum cell thickness of lopped cells will be 100.
This, I think, is a much cleaner specification. AJA.